+++ ample-runtime/src/main/java/app/revanced/manager/patcher/ample/AmpleSession.kt
@@ -316,6 +316,12 @@ class AmpleSession private constructor(
                         message = "Applying patched changes"
                     )
                 )
+                onEvent(
+                    ProgressEvent.Progress(
+                        stepId = StepId.WriteAPK,
+                        message = "Compiling modified resources"
+                    )
+                )
                 runCancellableBlockingIo(checkCancelled) {
                     result.applyTo(patched)
                 }